pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-telescope"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "astropy>=6.0",
    "numpy>=1.26",
    "pillow>=10.0",
]

[project.scripts]
smarttscope = "smart_telescope.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.15",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.10",
    "pyserial>=3.5",
    "build>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["smart_telescope*"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
asyncio_mode = "auto"
addopts = "--cov=smart_telescope --cov-report=term-missing --cov-report=xml --cov-fail-under=80"
